Renamed all files to lower-case.
