+++ src/core/component/logic.js
