+++ src/core/entity.js
