+++ src/core/resource.js
