+++ src/core/scene.js
