+++ src/core/thread.js
